Move instrumentation configs to global InstrumentationOptions
Move OpenAI Agents, LangChain, and Microsoft Agent Framework instrumentation
options from A365Options.instrumentations into the root-level
InstrumentationOptions interface in src/types.ts. This unifies all
instrumentation toggles at the distro level rather than scoping them
under A365.

- Added openaiAgents, langchain, microsoftAgentFramework to InstrumentationOptions
- Moved OpenAIAgentsInstrumentationConfig and LangChainInstrumentationConfig to src/types.ts
- Removed A365InstrumentationOptions and instrumentations from A365Options
- Updated barrel exports and tests
