+++ src/types.ts
@@ -49,6 +49,53 @@ export interface InstrumentationOptions {
   bunyan?: InstrumentationConfig;
   /** Winston Instrumentation Config */
   winston?: InstrumentationConfig;
+
+  // ── GenAI & agent framework instrumentations ──────────────────────
+
+  /**
+   * OpenAI Agents SDK instrumentation.
+   * Pass `true` for defaults or a configuration object.
+   * Requires `@openai/agents` as an optional peer dependency.
+   */
+  openaiAgents?: boolean | OpenAIAgentsInstrumentationConfig;
+
+  /**
+   * LangChain instrumentation.
+   * Pass `true` for defaults or a configuration object.
+   * Requires `@langchain/core` as an optional peer dependency.
+   */
+  langchain?: boolean | LangChainInstrumentationConfig;
+
+  /**
+   * Microsoft Agent Framework instrumentation.
+   * Requires `@microsoft/agents-hosting` as an optional peer dependency.
+   */
+  microsoftAgentFramework?: boolean;
+}
+
+/** Configuration for OpenAI Agents SDK instrumentation. */
+export interface OpenAIAgentsInstrumentationConfig {
+  /** Custom tracer name. */
+  tracerName?: string;
+  /** Custom tracer version. */
+  tracerVersion?: string;
+  /**
+   * When true, the gen_ai.input.messages attribute will be suppressed
+   * on InvokeAgent scope spans.
+   * @default false
+   */
+  suppressInvokeAgentInput?: boolean;
+  /**
+   * Enable recording of message content (input/output messages, tool args, etc.) in spans.
+   * @default false
+   */
+  isContentRecordingEnabled?: boolean;
+}
+
+/** Configuration for LangChain instrumentation. */
+export interface LangChainInstrumentationConfig {
+  /** Enable recording of message content in spans. */
+  isContentRecordingEnabled?: boolean;
 }
 
 /**
